Replace using a class property with an interface method
Class BaseDNSBLClient was designed to use _LISTS_ATTR_NAME class
property to investigate its dependencies, expected to be dnsbl service objects. Each
such object was tested for both having an attribute with a name equal to value
of _LISTS_ATTR_NAME and set to true. The actual value of _LISTS_ATTR_NAME was
to be set in subclasses of BaseDNSBLClient, or, in tests, to an arbitrary value.

However, this implementation was problematic. Avoiding simple copying of names
of lists_* attributes of DNSBLService and setting them as string values to
_LISTS_ATTR_NAME required two things:

* modification of DNSBLService so it would store names of these properties
as strings, and acces their values using these strings
* introducing dependency on DNSBLService class to subclasses of BaseDNSBLClient

This is solved by removing _LISTS_ATTR_NAME and replacing it with _required_data_in
method that must be implemented in the subclasses.