Remove BaseDNSBLClient and related classes
BaseDNSBLClient, DNSBLClient and URIDNSBLClient classes will not
be used. The task of preparing queried data once for many instances of DNSBLService
class will be performed by an object whose only responsibility is memoization.

Implementation of DNSBLService will be changed so that it could query the service
it represents for varous initial data: host strings, valid ip address objects and valid
domain name objects.