Reorganize test for DNSBLService
Due to changes in the architecture of the project, the DNSBLServiceTest class
is being reorganized. It's name is changed to BaseDNSBLTest to reflect new name
of DNSBLService class, and its data and methods are reorganized for testing
new methods designed for BaseDNSBL and its subclasses.

First, data is being made available in form of both strings and instances
of ip objects (from ipaddress module) and dns name objects (from dns module),
so that methods receiving various types of these arguments could be tested.

Second, a new methods starting with "do-" are added as generalizations of
tests for some methods of BaseDNSBL class: doTestContains, doTestContainsHostname,
doTestContainsIp, doTestNotImplementedOperation.

Third, tests for all methods of BaseDNSBL object are added to BaseDNSBLTest
class, but all of them test expected result of these methods raising NotImplementedError,
due to the fact that BaseDNSBL is designed with it's basic methods responsible
for querying ip address and dns name objects to be implemented in separate classes.