Remove lists_ips and lists_uris properties from BaseDNSBL
These values are not needed for the new architecture: ability to query for ip addresses
or for hostnames will be based on a subclass of BaseDNSBL having implementation
of proper methods.

The code responsible for querying dnsbl services combined with other sources
will use EAFP coding style, so it also won't need an interface to make sure the
right sources are queried for given value.