Reimplement BaseDNSBL._query method
The method is being reimplemented according to the new design. The new
method will accept string values and either use them to create ip address
objects or relative dns name objects, and then pass them to other methods, or
raise an exception if the value was not a valid host.