Reorganize tests for DNSBL subclasses
The code responsible for testing methods of BaseDNSBL class and its subclasses
is being moved to separate classes:

* DNSBLInvalidValueTestMixin
* BaseDNSBLValidArgumentTestMixin
* DNSBLIpTestMixin
* DNSBLDomainTestMixin

The code left in BaseDNSBLTest is responsible for preparations before running
test methods. This class is being temporarily moved to AbstractTests class to prevent
it from being instantiated and used by automatic tests.

The instantiation of objects of tested classes and the actual tests are now being
performed in the following classes:

* IpDNSBLTest
* DomainDNSBLTest
* GeneralDNSBLTest

Tests for host-type specific methods (lookup_ip, query_for_hostname, etc.) are also
being removed.