Reorganize tests for subclasses of BaseDNSBL
Subclasses of BaseDNSBL have two tested methods:
__contains__ and lookup. These methods are expected to return different values
or raise different exceptions based on membership of their arguments in
the following groups:

* listed hosts
* not listed hosts
* listed hosts with return codes not included in classification map passed to
constructor of instance of BaseDNSBL subclasses
* invalid arguments (not recognized as hosts acceptable by instance of BaseDNSBL
subclass)

The difference between subclasses of BaseDNSBL is in types of hosts they accept as valid.
This reorganization of tests utilizes it. Instead of relying on inheritance and overriding
test methods provided by mixins used in different combinations, tests can be
constructed as generalized for the all subclasses. The new tests use arguments
classified in the provided groups. They are implemented in BaseDNSBLTest class,
which is no longer nested in another class - instead, to prevent its direct use during
testing, it simply extends the object class.

The concrete test classes extend both BaseDNSBLTest and unittest.TestCase, and they
provide ip addresses and hostnames as arguments to be used during testing.
The arguments are properly classified for subclasses of BaseDNSBL being tested.