Replace argument of RedirectUrlResolver constructor
The original argument, max_retries, has been passed to a session object -
an instance of requests.Session. Now, it is being replaced with an instance of this class.
This way, not only max_retitries, but all configuration options that are available
for requests.Session become available to user of RedirectUrlResolver.