Modify tests of BaseUrlTester.urls_to_test method
The modifications take into account the expectations of urls_to_test method:
the actual expected order of each item in expected sequence is no longer tested.
Instead, the membership of items in the result sequence is tested, and the requirement
of urls resulting from redirect resolution always following initial values in the result
sequence.