Refactor BaseUrlTesterTest
The purpose of refactoring is to make the test case use Url named tuple and
all methods added to the class in previous commits.

It makes some methods unnecessary, so they are also subject to
removal in this refactoring. The methods are:
* get_location
* get_redirect_slice

Some test methods were also removed. The new method of handling redirect
histories (which are no longer stored in separate lists and grouped by
the type of target location) and introduction of getExpectedRedirectUrls method,
allowed to generalize tests involving different kinds of final target locations.