Refactor code responsible for adding test data
Specific methods for adding redirects with invalid, http and non-http targets
to test data for BaseUrlTesterTest are not necessary. The methods are being removed,
and the method used to implement them is used instead. Also, it is being renamed and
the return value is being removed - it's better to maintain command-query
separation than to have a method that could be reused perhaps just once.