Improve tests for spambl.IpAddress
The improvements include:
* removing unnecessary comments
* reorganizing test methods using parameterization
* adding two-way tests for is_match method of two instances of IpAddress, each
test is now performed by calling the method on one of the objects and passing the
other as argument, for both combinations of objects.

Such a test makes sense for another subclass of Host: Hostname. In both combinations,
the results could be different. Therefore, it's good to test this method in this way also
for instances of IpAddress class.

Perhaps the name of the method is_match does not reflect that possiblity very well, in which case it will
probably be renamed.