Move some common valid url tests to a separate class
BaseUrlTesterTest has the following attributes:

* test_any_match_returns_true_for and test_lookup_matching_for -
methods decorated with parameterized.expand decorator. They serve
as bases for parameterized methods generated by the decorator
* valid_url_input and valid_url_list_input  - used as input parameters
for the test method generation

The test methods generated from them are common for several test case classes
for url testers, but not for all of them. For some, they are not aplicable and should
be replaced with their equivalents which differ at least in input parameters. Therefore,
they are being moved into a separate class: CommonValidUrlTest.