Add type conversion to list for an argument
matching_urls, an argument of two methods of BaseUrlTesterTest:
(_test_any_match_returns_true and _set_matching_urls), is used in
both methods by adding it to a sequence, assuming both sequences are lists.
To be able to reuse these methods with matching_urls being a dict in which
matching urls are its keys, a type conversion to list is being added.