Refactor base url tester test classes
The classes being refactored are BaseUrlTesterTest and CommonValidUrlTesterTest
A new class is being created: GeneratedUrlTesterTest, and test methods generated
with parameterized.expand are being moved to this class from BaseUrlTesterTest and
CommonValidUrlTesterTest, with appropriate data necessary for the test generation.

BaseUrlTesterTest is left with only non-generated test methods and methods
serving as bases for tests for some url tester methods, while CommonValidUrlTesterTest
is being removed.

All test case classes using BaseUrlTesterTest and CommonValidUrlTesterTest are
being refactored to acommodate for this change.