Generalize TestFunctionDoesNotHandleProvider for generators
Now the value returned by the tested function is tested for being instance
of types.GeneratorType. If the test returns True, a list is created with
the result as argument, to cover all the code expected to be covered in
generator. This code is expected to raise an error of given type.