Make accepts_valid_host decorator be used for host validation
Using accepts_valid_host is a new validation method that will replace
relying on exceptions raised by host factory used by an instance of HostList
subclass.

The decorated methods are: __contains__ and lookup of HostList class.

HostCollection.add method is not being decorated. An error resulting from
an attempt to add a host unsupported by the host factory used by the class
should not be ingored, and since it won't be, using accepts_valid_host is redundant.