Make __contains__ and lookup methods handle InvalidHostError
From now, the InvalidHostError raised by host factory of a HostList subclass
instance is being handled by __contains__ and lookup methods by returning
a value expected for hosts not listed by the service being queried.

This change is introduced to make assembling composite url tester
(with UrlTesterChain) more practical - before it, querying for a set of urls
could be interrupted by querying a service for a host it does not support.