Refactor IPv6 handling code in HpHosts
hpHosts does not support IPv6 hosts, which was reflected in HpHosts class
by raising an error in response to querying for an IPv6 value. This commit
changes it by removing the code responsible for testing the host value
and raising the error and replacing it with using a new function as host factory -
non_ipv6_host. This way, hpHosts service is still not queried for IPv6 values,
and using the client for querying hpHosts for such a value results in
a proper value being returned by __contains__ and lookup methods.