Move pylint-related block comments
This commit changes placement of the block comments so that they are placed
at the end of lines or inside code blocks to which they refer.

Previously, I thought a pylint block comment added before a line on which
pylint information appeared would work as a kind of annotation, disabling
the message placed on the next line or refering to the next block of code.
It turns out I was wrong.

This probably also explains all previous incidents with pylint problems
dissapearing in both code and eclipse problem view, and then reappearing.
As in this case, this dissapearance and reappearing was probably a result
of adding module-level pylint block comments and then removing them with
the code to which they were supposed to refer.