Change the way HostCollection.hosts property is initialized
This commit changes implementation of __init__ method, so that
instead of setting HostCollection.hosts property to set and
then using HostCollection.add method to populate it, the __init__
method simply assigns the value of hosts argument to the property, or,
if hosts is None, to a new instance of set.

The default value of hosts argument is being changed from an empty
tuple to None, and non-default value is expected to have .add and .remove
methods, instead of just being an iterable.