Change expected interface and default value of HostCollection.hosts
The default value is being changed from set to list, and expected interface:
from one containing methods of a set object to one containing methods
of a list object.

The reasoning behind this change is that constraints for values stored in
a host collection are more than just having unique values: each hostname
stored in the collection must cover a domain not covered by another hostname
(that is: when host colletion contains a domain, items explicitly representing
its subdomains can not be listed), and that is already handled by the code
of the collection, which makes set unnecessary when another type could
be used.

Another reason for this change is that it will allow better code
restructuring for introduction of another class of host collection,
storing items in specific order and using binary search. Common
interface for objects storing the host values in both classes means
introducing a common base class for host collections would be simpler:
more code could be shared between the classes, and we wouldn't have
to handle different types of  default values of .hosts property for
each of the classes.

In addition, alternative classes to be used for .hosts will have to
have a simpler interface to be used by both unsorted and sorted
host collections.