Change the way Hostname class uses dns.name.Name class
This commit removes dns.name.Name from base classes of Hostname class
and modifies the code of the Hostname class so that it stores instance
of dns.name.Name. This way, no method of the original class is overridden,
and its instance is always available to the user of the library.