Change the way subclasses of IPAddress use classes from ipaddress module
This commit modifies IPAddress class and its subclasses. First, it makes
IPAddress use a constructor (to be specified in subclasses) for creating
an instance of ip address object, which is then stored in instance
of a subclass of IPAddress. Then, it modifies subclasses of IPAddress so that
they specify ipaddress.IPv4Address and ipaddress.IPv6Address as classes
of objects they use as their values.

This way, no method of original classes is overridden, and their instances
are available to the user of the library.