Refactor test.unit.test_service_models and test_host_collections
The test modules are being refactored by moving definitions of:
* UrlTesterTestMixin
* get_hosts
* host_list_host_factory
* HostListTestMixin

from test.unit.test_service_models to test.unit.common_definitions,
and modifying imports in both test_service_models and test_host_collections
to accommodate for the change.

The classes and functions moved are either dependencies of one of
the moved classes (in case of get_hosts), or are to be shared between
multiple modules.