Move definitions of client classes (and their dependencies) to clients module
This commit moves definitions of DNSBL, BitmaskingDNSBL, HpHosts
and GoogleSafeBrowsing classes, and get_powers_of_2 function, from
spam_lists.service_models module to spam_lists.clients module.

I think it is better to put classes of clients and some preconfigured
instances of these classes provided by this library in the same module.

After moving classes of custom host collections to
spam_lists.host_collections, and classes of service clients to
spam_lists.clients, the only class left at the old service_models module
is HostList. This class is neither a class of client objects, nor a class of
objects representing a custom host collection, but it is their common base.
As such, it should be stored in a separate module and imported in both
host_collections and clients modules.