Replace overlooked ocurences of the old name of the library with the new one
Occurences of 'spambl' - the old name of the library - have been overlooked
in unit tests for HpHosts client. This commit replaces them with "spam_bl" -
the new name of the library, with underscore used as a replacement for
hyphen, to match the convention of values of client_name argument
of HpHosts constructor, used in the tests.