Fix tests of Host.__lt__ method for Python 3
This commit refers to tests for beahviour of __lt__ method when comparing
two host objects with uncomparable values of their .value attributes. In
such situation (which occurs, for example, when comparing Hostname and
IPv4Address objects, which store instances of uncomparable types dns.name.Name
and ipaddress.IPv4Address, respectively), the method is expected to compare
return values of to_unicode method of both objects. This turns out to be
problematic when running tests with Python 3 interpreter, for two reasons:

* the initial implementation of the test contains an error: it assumes
the method uses return value of self.value.to_unicode() call, instead of
that of self.to_unicode(). This leads to TypeError resulting from comparing
instances of MagicMock and unicode object, which aren't comparable in
Python 3.
* fixing the first error exposes another one: subclasses of
spam_lists.structures.IPAddress use call to str(self.value) as their
implementation of to_unicode method, which always returnes a string object
during tests. This makes the original implementation (relying on comparison of
return values of to_unicode methods) of the test pass for these classes, while
it fails for Hostname class, because it uses call to self.value.to_unicode
as its implementation for to_unicode method. This call returns an instance
of MagicMock, which is not comparable with a string in Python 3.

Providing mock implementation of __lt__ method for MagicMock instance
fixes the error for the test for Hostname class, but introduces another
error for tests for IPAddress subclasses, because in their case, to_unicode
returns a string, not a MagicMock, and .__lt__ method of a string obviously
doesn't have the return_value property of a MagicMock instance.

The solution applied by this commit relies on removing the original
test for __lt__ method reacting to uncomparable values, and providing
separate implementations of the test for both Hostname and IPAddress
subclasses.