Remove inconsistency in behaviour of Host.__lt__ method
The implementation of the method relies on comparison between self.value
and other.value, which is implemented by self.value.__lt__ method. This
method may return a NotImplemented constant. Because until now, Host.__lt__
performed the comparison using "<" operator, the NotImplemented value was
automatically handled by Python interpreter. Moreover, when comparing the self
to the other object without either value or to_unicode attributes, Host.__lt__
itself returned NotImplemented value, which was also to be handled automatically
if the method was called as a result of evaluating an expression with the "<"
operator.

However, the way Python handles the value differs between the two major
versions of the language: in Python 2, comparison is deffered elsewhere and
handled silently, resulting in a boolean value being returned, while in
Python 3, it is handled by raising TypeError.

This commit removes this inconsistency by introducing the following changes:
* Host.__lt__ directly calls self.value.__lt__ and handles its NotImplemented
return value by comparing unicode string values of both self and the other
* AttributeError is being handled by raising TypeError