Disable W0104 warning in a test of a property
W0104 is a pylint warning about a statement that seems to have no
effect. Here, the statement is accessing a tested property so that
the underlying function can be tested. The value of the property is not
used in the test so, instead of assigning it to an unused variable, I
decided to just disable the warning.